# contracts/openapi/psma.openapi.json
{
  "components": {
    "schemas": {
      "Attribution": {
        "properties": {
          "required": {
            "default": true,
            "title": "Required",
            "type": "boolean"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "text"
        ],
        "title": "Attribution",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ProviderEnvelope": {
        "properties": {
          "attribution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Attribution"
              },
              {
                "type": "null"
              }
            ]
          },
          "data": {
            "title": "Data"
          },
          "provider": {
            "description": "Provider identifier (e.g. tvmaze, tmdb)",
            "title": "Provider",
            "type": "string"
          },
          "request": {
            "additionalProperties": true,
            "title": "Request",
            "type": "object"
          },
          "retrieved_at": {
            "description": "When PSMA fetched this data (UTC)",
            "format": "date-time",
            "title": "Retrieved At",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "data"
        ],
        "title": "ProviderEnvelope",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Program Subscription Manager Application (PSMA) backend API",
    "title": "PSMA API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/providers/tmdb/search/tv": {
      "get": {
        "operationId": "tmdb_search_tv_providers_tmdb_search_tv_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language"
            }
          },
          {
            "in": "query",
            "name": "include_adult",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Adult",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tmdb Search Tv",
        "tags": [
          "providers"
        ]
      }
    },
    "/providers/tmdb/tv/{series_id}/watch/providers": {
      "get": {
        "operationId": "tmdb_tv_watch_providers_providers_tmdb_tv__series_id__watch_providers_get",
        "parameters": [
          {
            "in": "path",
            "name": "series_id",
            "required": true,
            "schema": {
              "title": "Series Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tmdb Tv Watch Providers",
        "tags": [
          "providers"
        ]
      }
    },
    "/providers/tvmaze/search/shows": {
      "get": {
        "operationId": "tvmaze_search_shows_providers_tvmaze_search_shows_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tvmaze Search Shows",
        "tags": [
          "providers"
        ]
      }
    },
    "/providers/tvmaze/shows/{show_id}": {
      "get": {
        "operationId": "tvmaze_get_show_providers_tvmaze_shows__show_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "show_id",
            "required": true,
            "schema": {
              "title": "Show Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "embed",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "cast",
                    "crew",
                    "episodes",
                    "seasons",
                    "akas",
                    "images",
                    "nextepisode",
                    "previousepisode"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Embed"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tvmaze Get Show",
        "tags": [
          "providers"
        ]
      }
    },
    "/version": {
      "get": {
        "operationId": "version_version_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Version Version Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Version"
      }
    }
  }
}